refactor: Nest public query-interceptor contract under the jaegerquery extension
The contract is jaeger_query's read-path SPI, not a standalone extension, so its
public re-export belongs under components/extension/jaegerquery/queryinterceptor
(mirroring the internal cmd/jaeger/internal/extension/jaegerquery/queryinterceptor)
rather than as a sibling of the jaegerquery extension.
